add astype

+++ log_reg_ilc_jax_sean.py
@@ -369,14 +369,14 @@ if __name__ == "__main__":
     for i in range(len(X_1)):          
         try:
             index = int(X_1[i].decode("utf-8").split('/')[-1].split('.')[0].split('_')[-1])
-            V1['X'][index-1] = np.load(X_1[i], mmap_mode='r')
+            V1['X'][index-1] = np.load(X_1[i], mmap_mode='r').astype(np.float32)
         except:
             print(f"An exception occurred at index {i} of V1: {X_1[i]}")
             continue
 
         try:
             index = int(y_1[i].decode("utf-8").split('/')[-1].split('.')[0].split('_')[-1])        
-            V1['y'][index-1] = np.load(y_1[i], mmap_mode='r')
+            V1['y'][index-1] = np.load(y_1[i], mmap_mode='r').astype(np.float32)
         except:
             print(f"An exception occurred at index {i} of V1: {y_1[i]}")
             continue
@@ -386,13 +386,13 @@ if __name__ == "__main__":
     for i in range(len(X_2)):     
         try:
             index = int(X_2[i].decode("utf-8").split('/')[-1].split('.')[0].split('_')[-1])
-            V2['X'][index-1] = np.load(X_2[i], mmap_mode='r')
+            V2['X'][index-1] = np.load(X_2[i], mmap_mode='r').astype(np.float32)
         except:
              print(f"An exception occurred at index {i} of V2: {X_2[i]}")
              continue
         try:
             index = int(y_2[i].decode("utf-8").split('/')[-1].split('.')[0].split('_')[-1])
-            V2['y'][index-1] = np.load(y_2[i], mmap_mode='r')
+            V2['y'][index-1] = np.load(y_2[i], mmap_mode='r').astype(np.float32)
         except:
             print(f"An exception occurred at index {i} of V2: {y_2[i]}")
             continue
@@ -402,14 +402,14 @@ if __name__ == "__main__":
     for i in range(len(X_3)):
         try:
             index = int(X_3[i].decode("utf-8").split('/')[-1].split('.')[0].split('_')[-1])
-            V3['X'][index-1] = np.load(X_3[i], mmap_mode='r')
+            V3['X'][index-1] = np.load(X_3[i], mmap_mode='r').astype(np.float32)
         except:
             print(f"An exception occurred at index {i} if V3: {X_3[i]}")
             continue
 
         try:
             index = int(y_3[i].decode("utf-8").split('/')[-1].split('.')[0].split('_')[-1])
-            V3['y'][index-1] = np.load(y_3[i], mmap_mode='r')
+            V3['y'][index-1] = np.load(y_3[i], mmap_mode='r').astype(np.float32)
         except:
             print(f"An exception occurred at index {i} if V3: {y_3[i]}")
             continue
